Move to scalar_cast_extension, need C++ tests [skip ci]
