+++ cpp/src/arrow/compute/kernels/scalar_cast_numeric.cc
@@ -16,16 +16,15 @@
 // under the License.
 
 // Implementation of casting to integer, floating point, or decimal types
+
 #include "arrow/array/builder_primitive.h"
 #include "arrow/compute/kernels/common.h"
 #include "arrow/compute/kernels/scalar_cast_internal.h"
 #include "arrow/compute/kernels/util_internal.h"
-#include "arrow/extension_type.h"
 #include "arrow/scalar.h"
 #include "arrow/util/bit_block_counter.h"
 #include "arrow/util/int_util.h"
 #include "arrow/util/value_parsing.h"
-#include "arrow/util/logging.h"
 
 namespace arrow {
 
@@ -769,36 +768,6 @@ std::vector<std::shared_ptr<CastFunction>> GetNumericCasts() {
 
   return functions;
 }
-
-Status CastToExtension(KernelContext* ctx, const ExecSpan& batch, ExecResult* out) {
-  const CastOptions& options = checked_cast<const CastState*>(ctx->state())->options;
-  auto out_ty = static_cast<const ExtensionType&>(*options.to_type.type).storage_type();
-  DCHECK(batch[0].is_array());
-  std::shared_ptr<Array> array = batch[0].array.ToArray();
-  std::shared_ptr<Array> result;
-  RETURN_NOT_OK(Cast(*array, out_ty, options,
-                     ctx->exec_context())
-                    .Value(&result));
-  
-  ExtensionArray extension(options.to_type.GetSharedPtr(), result);
-  out->value = std::move(extension.data());
-  return Status::OK();
-}
-
-std::shared_ptr<CastFunction> GetCastToExtension(std::string name) {
-  auto func = std::make_shared<CastFunction>(std::move(name), Type::EXTENSION);
-  for (auto in_ty : IntTypes()) {
-      DCHECK_OK(func->AddKernel(in_ty->id(), {in_ty}, 
-                                kOutputTargetType, CastToExtension));
-  }
-  return func;
-}
-
-std::vector<std::shared_ptr<CastFunction>> GetExtensionCasts() {
-  auto func = GetCastToExtension("cast_extension");
-  return {func};
-}
-
 
 }  // namespace internal
 }  // namespace compute
